+++ tests/testthat/test-disambiguate_terms.R
@@ -20,11 +20,8 @@ test_that("tidy_disambiguate_terms() works for mixed models", {
     tidy_disambiguate_terms(sep = ".")
   expect_equal(
     res$term,
-    c(
-      "(Intercept)", "Days", "Subject.sd__(Intercept)",
-      "Subject.cor__(Intercept).Days",
-      "Subject.sd__Days", "Residual.sd__Observation"
-    )
+    c("(Intercept)", "Days", "Subject.sd__(Intercept)", "Subject.sd__Days",
+      "Subject.cor__(Intercept).Days", "Residual.sd__Observation")
   )
   expect_true("original_term" %in% names(res))
 
@@ -33,11 +30,8 @@ test_that("tidy_disambiguate_terms() works for mixed models", {
     tidy_disambiguate_terms(sep = "_")
   expect_equal(
     res$term,
-    c(
-      "(Intercept)", "Days", "Subject_sd__(Intercept)",
-      "Subject_cor__(Intercept).Days",
-      "Subject_sd__Days", "Residual_sd__Observation"
-    )
+    c("(Intercept)", "Days", "Subject_sd__(Intercept)", "Subject_sd__Days",
+      "Subject_cor__(Intercept).Days", "Residual_sd__Observation")
   )
 })