+++ NEWS.md
@@ -5,6 +5,11 @@
 - support for `survival::coxphms.object` models, see the
   experimental tidier `tidy_coxphms()` (#308)
 
+**Fixes**
+
+- `.clean_backticks()` fixed when variable names contain `$` 
+  (#311, @NourEdinDarwish)
+
 # broom.helpers 1.22.0
 
 **New supported models**